agent-service/ingest/requirements.txt
python-docx==1.1.2
python-pptx==1.0.2
qdrant-client==1.12.1
tiktoken==0.8.0
